added jarjarblinkz to launcher files
