+++ Launcher/App/src/main/java/com/jarjarblinkz/piLauncherNext/platforms/VRPlatform.java
@@ -1,4 +1,4 @@
-package com.veticia.piLauncherNext.platforms;
+package com.jarjarblinkz.piLauncherNext.platforms;
 
 import android.content.Context;
 import android.content.Intent;
@@ -7,16 +7,19 @@ import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.util.Log;
 
-import com.veticia.piLauncherNext.SettingsProvider;
+import com.jarjarblinkz.piLauncherNext.SettingsProvider;
 
 import java.util.ArrayList;
 
-public class AndroidPlatform extends AbstractPlatform {
+public class VRPlatform extends AbstractPlatform {
     public ArrayList<ApplicationInfo> getInstalledApps(Context context) {
         ArrayList<ApplicationInfo> installedApps = new ArrayList<>();
+        if (!isSupported()) {
+            return installedApps;
+        }
         PackageManager pm = context.getPackageManager();
         for (ApplicationInfo app : pm.getInstalledApplications(PackageManager.GET_META_DATA)) {
-            if (!isVirtualRealityApp(app)) {
+            if (isVirtualRealityApp(app)) {
                 if (!SettingsProvider.launchIntents.containsKey(app.packageName)) {
                     SettingsProvider.launchIntents.put(app.packageName, pm.getLaunchIntentForPackage(app.packageName));
                 }
@@ -38,16 +41,16 @@ public class AndroidPlatform extends AbstractPlatform {
         return installedApps;
     }
 
+    public boolean isSupported() {
+        return isMagicLeapHeadset() || isOculusHeadset() || isPicoHeadset();
+    }
+
     @Override
     public boolean runApp(Context context, ApplicationInfo app, boolean multiwindow) {
         Intent launchIntent = context.getPackageManager().getLaunchIntentForPackage(app.packageName);
         if (launchIntent != null) {
-            if (multiwindow) {
-                context.getApplicationContext().startActivity(launchIntent);
-            } else {
-                context.startActivity(launchIntent);
-            }
-        }else{
+            context.getApplicationContext().startActivity(launchIntent);
+        } else {
             Log.e("runApp", "Failed to launch");
             return false;
         }
