+++ Launcher/App/src/main/java/com/jarjarblinkz/piLauncherNext/ui/TextView.java
@@ -1,4 +1,4 @@
-package com.veticia.piLauncherNext.ui;
+package com.jarjarblinkz.piLauncherNext.ui;
 
 import android.content.Context;
 import android.content.res.TypedArray;
@@ -9,7 +9,7 @@ import android.graphics.drawable.Drawable;
 import android.text.TextPaint;
 import android.util.AttributeSet;
 
-import com.veticia.piLauncherNext.R;
+import com.jarjarblinkz.piLauncherNext.R;
 
 public class TextView extends androidx.appcompat.widget.AppCompatTextView {
 	private float strokeWidth;
