+++ Launcher/App/src/main/res/layout/lv_app.xml
@@ -30,7 +30,7 @@
             android:visibility="gone" />
     </RelativeLayout>
 
-    <com.veticia.piLauncherNext.ui.TextView
+    <com.jarjarblinkz.piLauncherNext.ui.TextView
         android:id="@+id/textLabel"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
